Make the Docker builds reproducible
In one year from now, if we were to clone this repository in a specific commit,
and build the Docker container, the resulting container must be the same as if
we built it today. This allow us to make sure the code that's running in
production is the same code that we tested locally. This patch fixes it.

The general ideia is to remove things that can change from our Dockerfile.
First, I created a `requirements.in` for `pip-compile` to generate a
`requirements.txt` file with all our dependencies specific versions. Then I did
the same for `package.json`, using `npm-shrinkwrap` to pin all versions down.

After that, I removed the `git clone`s from the Dockerfile. Before we were
cloning this repository and `os-data/eu-structural-funds` when building the
container. Instead, we now add the files from the current repository. To make
it work with `eu-structural-funds` as well, I had to add it as a `git-subtree`
in this repository. Whenever we want to update the data, we can run `make
update-eu-structural-funds`.

All in all that was it. The rest of the changes are just minor refactorings.